Update to latest milestone changes (#927)

+++ src/client.rs
@@ -919,11 +919,11 @@ impl Client {
     }
 
     /// Returns checked local time and milestone index.
-    pub async fn get_time_and_milestone_checked(&self) -> Result<(u64, u32)> {
+    pub async fn get_time_and_milestone_checked(&self) -> Result<(u32, u32)> {
         let local_time = SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .expect("Time went backwards")
-            .as_secs();
+            .as_secs() as u32;
         let status_response = self.get_info().await?.nodeinfo.status;
         let latest_ms_timestamp = status_response.latest_milestone_timestamp;
         // Check the local time is in the range of +-5 minutes of the node to prevent locking funds by accident
